+++ packages/cli-harnesses/harness-default-to-non-default/Cargo.toml
@@ -13,3 +13,4 @@ dioxus = { workspace = true, features = [] }
 default=["web"]
 web=["dioxus/web"]
 
+
