+++ packages/cli-harnesses/harness-fullstack-multi-target/Cargo.toml
@@ -16,3 +16,4 @@ desktop=["dioxus/desktop"]
 mobile=["dioxus/mobile"]
 server=["dioxus/server"]
 
+
