+++ packages/cli-harnesses/harness-fullstack-with-optional-tokio/Cargo.toml
@@ -16,3 +16,4 @@ default = []
 server = ["dioxus/server", "dep:tokio"]
 web = ["dioxus/web"]
 
+
